Retype filed CountryHouse

+++ CountryHouse.cpp
@@ -1,6 +1,6 @@
 #include "CountryHouse.h"
 
-CountryHouse::CountryHouse(int worth, double distanceFromCity): Property(worth)
+CountryHouse::CountryHouse(int worth, unsigned int distanceFromCity): Property(worth)
 {
 	this->distanceFromCity = distanceFromCity;
 }
